refactor tests of federated query response into separate module
